Open an image, and print the dimensions.
Right now, we're going to panic when the file doesn't exist, or isn't a
supported image.

In the future, we don't want the software to panic when the file isn't a
supported image.  We'd like to skip processing that file in that case.

We only want to panic if we're being given file names that don't exist.