Use new processor module.
Also examine how we're going to determine what the image's format
actually is.  This will be needed if we are going to fix the extension.

We cannot use image::open(), since in this approach, the image library
will assume the format from the file's extension.  We don't want this.

We'll use a Reader, and use with_guessed_format() to read the image
header to get the actual format.