Implement dry-run, add output.
Enable recursive mode.  Software runs in recursive mode by default if
given a directory.

Create process() that takes a path, then determines if it is a file or
directory, and calls the appropriate function as needed.

Produce cleaner output.

Get rid of debug prints.

Early exit in the case of unsupported or unrecognized formats.  We don't
want the software to panic in either of these cases.  It should skip
over the file and go to the next file.