Handle decode errors.
Don't panic here, let's take a look at the errors we get when decoding
image data.

As a side note, profiling shows that the decode step takes the most
time.  There are some errors that can pop up at the decode stage, and it
may be nice to catch them.