Add a decode flag.
Add the option to decode image data.

Originally, the software would open and decode image data.  However, it
was shown that the decode step took quite a bit of time.

We may want to catch decoding errors as well to track down invalid image
files.  A decode flag has been added that allows the user to check for
decode errors.

This change results in a significant speed up when not decoding image
data.

The default is not to decode image data.